requirements.txt
Django==4.2.5
python-dotenv==1.0.0
uagents==0.6.2

